refactor: id 필드 타입 변경

+++ src/main/java/com/example/community/dto/comment/CommentCreateRequestDto.java
@@ -10,7 +10,7 @@ import lombok.Setter;
 @Setter
 @AllArgsConstructor
 public class CommentCreateRequestDto {
-    private Integer userId;
-    private Integer postId;
+    private Long userId;
+    private Long postId;
     private String contents;
 }
